// src/layout/types.ts
export type DisplayValue =
  | "block"
  | "flex"
  | "initial"
  | "inline"
  | "inline-block"
  | "inline-flex"
  | "none"
  | "hide-accessible"; // TODO: https://jira.xarth.tv/browse/COREUI-3554
// ...
export enum ZIndex {
  Default = "z-index-default",
  Above = "z-index-above",
  Below = "z-index-below",
}

// `ZIndexValue` is not defined to avoid possible confusion; only the Enum is accepted for now.

export type FlexValue = 0 | 1 | 2 | 3 | 4;

// from inject-styled-layout

export enum Background {
  Base = "base",
  Alt = "alt",
  Alt2 = "alt-2",
  Overlay = "overlay",
  Accent = "accent",
  AccentAlt = "accent-alt",
  AccentAlt2 = "accent-alt-2",
  Inherit = "inherit",
}

export type BackgroundValue =
  | "base"
  | "alt"
  | "alt-2"
  | "overlay"
  | "accent"
  | "accent-alt"
  | "accent-alt-2"
  | "inherit"; // TODO: https://jira.xarth.tv/browse/COREUI-3554

export enum Color {
  Base = "base",
  Alt = "alt",
  Alt2 = "alt-2",
  Link = "link",
  Live = "live",
  Error = "error",
  Overlay = "overlay",
  OverlayAlt = "overlay-alt",
  Inherit = "inherit",
}

export type ColorValue =
  | "base"
  | "alt"
  | "alt-2"
  | "link"
  | "live"
  | "error"
  | "overlay"
  | "overlay-alt"
  | "inherit"; // TODO: https://jira.xarth.tv/browse/COREUI-3554

export type Elevation = 0 | 1 | 2 | 3 | 4 | 5;

export enum FontSize {
  Size1 = "font-size-1",
  Size2 = "font-size-2",
  Size3 = "font-size-3",
  Size4 = "font-size-4",
  Size5 = "font-size-5",
  Size6 = "font-size-6",
  Size7 = "font-size-7",
  Size8 = "font-size-8",
}

// `FontSizeValue` is not defined to avoid possible confusion; only the Enum is accepted for now.

export enum FontWeight {
  Regular = "regular",
  SemiBold = "semibold",
  Bold = "bold",
}

export type FontWeightValue = "regular" | "semibold" | "bold"; // TODO: https://jira.xarth.tv/browse/COREUI-3554

export type BorderRadiusProps = BorderRadius | BorderRadiusValues;